#60 Refactor after review
