+++ src/main/java/com/solvd/qaprotours/config/kafka/KafkaMessage.java
@@ -1,4 +1,4 @@
-package com.solvd.qaprotours.web.kafka;
+package com.solvd.qaprotours.config.kafka;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
