+++ src/main/java/com/solvd/qaprotours/config/kafka/KfProducerConfig.java
@@ -1,4 +1,4 @@
-package com.solvd.qaprotours.web.kafka;
+package com.solvd.qaprotours.config.kafka;
 
 import com.jcabi.xml.XML;
 import org.apache.kafka.clients.admin.AdminClientConfig;
