+++ src/main/java/com/solvd/qaprotours/config/kafka/LocalDateTimeDeserializer.java
@@ -1,4 +1,4 @@
-package com.solvd.qaprotours.web.kafka;
+package com.solvd.qaprotours.config.kafka;
 
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
