+++ src/main/java/com/solvd/qaprotours/config/kafka/elasticsearch/DeleteHandler.java
@@ -1,4 +1,4 @@
-package com.solvd.qaprotours.web.kafka.elasticsearch;
+package com.solvd.qaprotours.config.kafka.elasticsearch;
 
 import co.elastic.clients.elasticsearch.ElasticsearchClient;
 import co.elastic.clients.elasticsearch.core.DeleteRequest;
