+++ src/main/java/com/solvd/qaprotours/config/kafka/elasticsearch/package-info.java
@@ -1,4 +1,4 @@
 /**
  * This package contains handlers for Kafka to sent objects into ElasticSearch.
  */
-package com.solvd.qaprotours.web.kafka.elasticsearch;
+package com.solvd.qaprotours.config.kafka.elasticsearch;
