+++ src/main/java/com/solvd/qaprotours/config/kafka/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains classes for Kafka.
+ */
+package com.solvd.qaprotours.config.kafka;
