+++ src/main/java/com/solvd/qaprotours/service/impl/Scheduler.java
@@ -1,12 +1,12 @@
 package com.solvd.qaprotours.service.impl;
 
+import com.solvd.qaprotours.config.kafka.KafkaMessage;
+import com.solvd.qaprotours.config.kafka.MessageSenderImpl;
 import com.solvd.qaprotours.domain.MailData;
 import com.solvd.qaprotours.domain.MailType;
 import com.solvd.qaprotours.service.TicketService;
 import com.solvd.qaprotours.service.UserClient;
 import com.solvd.qaprotours.web.dto.MailDataDto;
-import com.solvd.qaprotours.web.kafka.KafkaMessage;
-import com.solvd.qaprotours.web.kafka.MessageSenderImpl;
 import com.solvd.qaprotours.web.mapper.MailDataMapper;
 import com.solvd.qaprotours.web.mapper.UserMapper;
 import lombok.RequiredArgsConstructor;
