--- src/main/java/com/solvd/qaprotours/web/kafka/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains classes for Kafka.
- */
-package com.solvd.qaprotours.web.kafka;
